% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{.check_data}
\alias{.check_data}
\title{Validity check}
\usage{
.check_data(input)
}
\arguments{
\item{input}{A numeric vector}
}
\value{
Nothing returned
}
\description{
This functions checks for NAs in a vector
}
