% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{.checkData}
\alias{.checkData}
\title{Validity check}
\usage{
.checkData(input)
}
\arguments{
\item{input}{A numeric vector}
}
\value{
Nothing returned
}
\description{
This functions checks for NAs in a vector
}
