% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{euclideanDist}
\alias{euclideanDist}
\title{Euclidean distance}
\usage{
euclideanDist(a, b, verbose = FALSE)
}
\arguments{
\item{a}{The first vector to use in the distance calculation.}

\item{b}{The second vector to use in the distance calculation.}

\item{verbose}{Boolean. If \code{TRUE}, a message will be printed.
Default \code{TRUE}.}
}
\value{
A numeric value of a distance
}
\description{
Calculates Euclidean distance between two vectors. An error will
be given if NAs are present in either vector.
}
\examples{
data(exampleData)
euclideanDist(exampleData[, 1], exampleData[, 2], verbose = FALSE)
}
