% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example dataset}
\format{
A matrix with 100 rows and 2 columns
}
\usage{
data("exampleData")
}
\description{
A dataset containing a matrix with two columns that were generated
with a random normal distribution with a mean of 0 and stdev of 1.
}
\examples{
data("exampleData")
}
\keyword{datasets}
