% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hello}
\alias{hello}
\title{Hello world}
\usage{
hello(withExcitement = TRUE)
}
\arguments{
\item{withExcitement}{Boolean. If \code{TRUE}, then the message
will be printed with exlamation points. Default \code{TRUE}.}
}
\value{
A string with the hello world text
}
\description{
A simple function to print Hello world
}
\examples{
hello(withExcitement = TRUE)
}
